# src/docs/swagger.yaml
definitions:
  accesstoken.AccessToken:
    properties:
      access_token:
        type: string
      client_id:
        type: integer
      expires:
        type: integer
      user_id:
        type: integer
    type: object
  accesstoken.AccessTokenRequest:
    properties:
      client_id:
        description: used for client credentials grant type
        type: string
      client_secret:
        type: string
      grant_type:
        type: string
      password:
        type: string
      scope:
        type: string
      username:
        description: used for password grant type
        type: string
    type: object
  oauth_errors.SwaggerRestErr:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
  /oauth/access_token:
    post:
      consumes:
      - application/json
      description: Generate and return new access token with provided information
      parameters:
      - description: Access Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/accesstoken.AccessTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/accesstoken.AccessToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/oauth_errors.SwaggerRestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/oauth_errors.SwaggerRestErr'
      summary: Create new access token
      tags:
      - access token
  /oauth/access_token/{access_token_id}:
    get:
      description: Return access token using access_token_id obtained via URL
      parameters:
      - description: Access Token ID
        in: path
        name: access_token_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/accesstoken.AccessToken'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/oauth_errors.SwaggerRestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/oauth_errors.SwaggerRestErr'
      summary: Get access token by its id
      tags:
      - access token
    patch:
      consumes:
      - application/json
      description: Set new expiration time, update db and return updated access token
      parameters:
      - description: Access Token ID
        in: path
        name: access_token_id
        required: true
        type: string
      - description: Access Token
        in: body
        name: at
        required: true
        schema:
          $ref: '#/definitions/accesstoken.AccessToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/accesstoken.AccessToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/oauth_errors.SwaggerRestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/oauth_errors.SwaggerRestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/oauth_errors.SwaggerRestErr'
      summary: Update given access token
      tags:
      - access token
swagger: "2.0"